Moved leia processors to a different module, allowing for custom implementations to be written on demand
