+++ leia-client-dropwizard/src/main/java/com/grookage/leia/dw/client/LeiaClientBundle.java
@@ -19,9 +19,9 @@ package com.grookage.leia.dw.client;
 import com.google.common.base.Preconditions;
 import com.grookage.leia.client.LeiaMessageProduceClient;
 import com.grookage.leia.client.datasource.NamespaceDataSource;
-import com.grookage.leia.client.processor.DefaultTargetValidator;
-import com.grookage.leia.client.processor.MessageProcessor;
-import com.grookage.leia.client.processor.TargetValidator;
+import com.grookage.leia.mux.processors.DefaultTargetValidator;
+import com.grookage.leia.mux.processors.MessageProcessor;
+import com.grookage.leia.mux.processors.TargetValidator;
 import com.grookage.leia.client.refresher.LeiaClientRefresher;
 import com.grookage.leia.client.refresher.LeiaClientSupplier;
 import com.grookage.leia.provider.config.LeiaHttpConfiguration;
