+++ leia-processors/pom.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
-  ~ Copyright (c) 2024. Koushik R <rkoushik.14@gmail.com>.
+  ~ Copyright (c) 2025. Koushik R <rkoushik.14@gmail.com>.
   ~
   ~ Licensed under the Apache License, Version 2.0 (the "License");
   ~ you may not use this file except in compliance with the License.
@@ -15,8 +15,8 @@
   ~ limitations under the License.
   -->
 
-<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xmlns="http://maven.apache.org/POM/4.0.0"
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
@@ -26,12 +26,7 @@
         <relativePath>../leia-parent</relativePath>
     </parent>
 
-    <artifactId>leia-client</artifactId>
-
-    <properties>
-        <lang3.version>3.11</lang3.version>
-        <maven.deploy.skip>false</maven.deploy.skip>
-    </properties>
+    <artifactId>leia-processors</artifactId>
 
     <dependencies>
         <dependency>
@@ -66,64 +61,12 @@
 
         <dependency>
             <groupId>com.grookage.leia</groupId>
-            <artifactId>leia-refresher</artifactId>
-        </dependency>
-
-        <dependency>
-            <groupId>com.grookage.leia</groupId>
-            <artifactId>leia-processors</artifactId>
-        </dependency>
-
-        <dependency>
-            <groupId>com.grookage.leia</groupId>
             <artifactId>leia-models</artifactId>
-        </dependency>
-
-        <dependency>
-            <groupId>com.grookage.leia</groupId>
-            <artifactId>leia-common</artifactId>
-        </dependency>
-
-        <dependency>
-            <groupId>com.grookage.leia</groupId>
-            <artifactId>leia-schema-validator</artifactId>
-        </dependency>
-
-        <dependency>
-            <groupId>com.jayway.jsonpath</groupId>
-            <artifactId>json-path</artifactId>
-            <version>2.9.0</version>
-        </dependency>
-
-        <dependency>
-            <artifactId>mockito-core</artifactId>
-            <groupId>org.mockito</groupId>
-            <scope>test</scope>
-        </dependency>
-
-        <dependency>
-            <artifactId>mockito-junit-jupiter</artifactId>
-            <groupId>org.mockito</groupId>
-            <scope>test</scope>
-        </dependency>
-
-        <dependency>
-            <artifactId>mockito-inline</artifactId>
-            <groupId>org.mockito</groupId>
-            <scope>test</scope>
-        </dependency>
-
-        <dependency>
-            <artifactId>leia-models</artifactId>
-            <type>test-jar</type>
-            <exclusions>
-                <exclusion>
-                    <artifactId>*</artifactId>
-                    <groupId>*</groupId>
-                </exclusion>
-            </exclusions>
-            <groupId>com.grookage.leia</groupId>
         </dependency>
     </dependencies>
+
+    <properties>
+        <maven.deploy.skip>true</maven.deploy.skip>
+    </properties>
 
 </project>
